requirements.txt
Django==3.1.5
djangorestframework==3.12.2
django_rest_swagger==2.2.0
coreapi==2.3.3
